Provide a way to set request variables in sync handler:
Note that Backbone.sync is a global function, and 
thus any overrides on the server have no access
to the current user. Therefore, the default 
backbone save/fetch methods have no way to be
properly scoped to the current user. This is a big
(and annoying) problem. The override here is for
ease of testing the backbone models.