Server side multi user backbone models, v0.0.16
Backbone models use a singleton for 
communication with the server. This works 
perfectly fine on the client (browser), where
only one user may be logged in. However, on the
server, it is possible that different backbone
model instances should belong to different users.

This commit extends both Model and Collection with
an 'init' static method. The first parameter
to init is an object that is used to extend the
request on the server.

Thus, to authenticate the request:

```js
var post = Post.init({user: user}, {_id: '23'});
post.fetch({success: success});
```