Add queryCache support: cache api requests on the server
This lets you avoid the double api request problem, where
a model or collection would be fetched when rendering on 
the server, and again when the app is initialized on the
client.

By outputting the queryCache into the server side rendered
page, and reinitializing the queryCache on app initialization,
that double fetch will be prevented, and the app initialization
will be sped up.